Move file with helper component for module category names
