+++ assets/locale/en/messages.pot
@@ -2148,43 +2148,43 @@ msgid "js.table.TableView.selectColumns.button"
 msgstr ""
 
 #. default-message: Add data
-#: assets/js/util/ModuleCategory.js:12
-msgid "js.util.ModuleCategory.CategoryNames.AddData"
+#: assets/js/ModuleCategoryName.js:12
+msgid "js.util.ModuleCategoryName.CategoryNames.AddData"
 msgstr ""
 
 #. default-message: Analyze
-#: assets/js/util/ModuleCategory.js:9
-msgid "js.util.ModuleCategory.CategoryNames.Analyze"
+#: assets/js/ModuleCategoryName.js:9
+msgid "js.util.ModuleCategoryName.CategoryNames.Analyze"
 msgstr ""
 
 #. default-message: Clean
-#: assets/js/util/ModuleCategory.js:8
-msgid "js.util.ModuleCategory.CategoryNames.Clean"
+#: assets/js/ModuleCategoryName.js:8
+msgid "js.util.ModuleCategoryName.CategoryNames.Clean"
 msgstr ""
 
 #. default-message: Code
-#: assets/js/util/ModuleCategory.js:11
-msgid "js.util.ModuleCategory.CategoryNames.Code"
+#: assets/js/ModuleCategoryName.js:11
+msgid "js.util.ModuleCategoryName.CategoryNames.Code"
 msgstr ""
 
 #. default-message: Combine
-#: assets/js/util/ModuleCategory.js:6
-msgid "js.util.ModuleCategory.CategoryNames.Combine"
+#: assets/js/ModuleCategoryName.js:6
+msgid "js.util.ModuleCategoryName.CategoryNames.Combine"
 msgstr ""
 
 #. default-message: Other
-#: assets/js/util/ModuleCategory.js:13
-msgid "js.util.ModuleCategory.CategoryNames.Other"
+#: assets/js/ModuleCategoryName.js:13
+msgid "js.util.ModuleCategoryName.CategoryNames.Other"
 msgstr ""
 
 #. default-message: Scrape
-#: assets/js/util/ModuleCategory.js:7
-msgid "js.util.ModuleCategory.CategoryNames.Scrape"
+#: assets/js/ModuleCategoryName.js:7
+msgid "js.util.ModuleCategoryName.CategoryNames.Scrape"
 msgstr ""
 
 #. default-message: Visualize
-#: assets/js/util/ModuleCategory.js:10
-msgid "js.util.ModuleCategory.CategoryNames.Visualize"
+#: assets/js/ModuleCategoryName.js:10
+msgid "js.util.ModuleCategoryName.CategoryNames.Visualize"
 msgstr ""
 
 #. default-message: {days}d ago
